config/requirements.txt
# Install CUDA-enabled torch first (recommended for RTX 4070 Ti Super):
# pip install torch torchvision torchaudio --index-url https://download.pytorch.org/whl/cu121

openai-whisper>=20240930
numpy>=1.26
tqdm>=4.66
